# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminInitiativeRow": {
        "properties": {
          "answer_count": {
            "title": "Answer Count",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "participant_type": {
            "title": "Participant Type",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "user_email": {
            "title": "User Email",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_email",
          "name",
          "participant_type",
          "status",
          "created_at",
          "answer_count"
        ],
        "title": "AdminInitiativeRow",
        "type": "object"
      },
...
      "AnswerCreate": {
        "properties": {
          "answer_value": {
            "$ref": "#/components/schemas/AnswerValue"
          },
          "followup_other": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Followup Other"
          },
          "followup_selections": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Followup Selections"
          },
          "mami_code": {
            "title": "Mami Code",
            "type": "string"
          },
          "question_id": {
            "title": "Question Id",
            "type": "string"
          },
          "questionnaire_version": {
            "title": "Questionnaire Version",
            "type": "string"
          }
        },
        "required": [
          "question_id",
          "mami_code",
          "questionnaire_version",
          "answer_value"
        ],
        "title": "AnswerCreate",
        "type": "object"
      },
      "AnswerRead": {
        "properties": {
          "answer_value": {
            "$ref": "#/components/schemas/AnswerValue"
          },
          "answered_at": {
            "format": "date-time",
            "title": "Answered At",
            "type": "string"
          },
          "followup_other": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Followup Other"
          },
          "followup_selections": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Followup Selections"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "initiative_id": {
            "title": "Initiative Id",
            "type": "integer"
          },
          "mami_code": {
            "title": "Mami Code",
            "type": "string"
          },
          "question_id": {
            "title": "Question Id",
            "type": "string"
          },
          "questionnaire_version": {
            "title": "Questionnaire Version",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "initiative_id",
          "question_id",
          "mami_code",
          "questionnaire_version",
          "answer_value",
          "followup_selections",
          "followup_other",
          "answered_at",
          "updated_at"
        ],
        "title": "AnswerRead",
        "type": "object"
      },
      "AnswerValue": {
        "enum": [
          "YES",
          "NOT_THERE_YET",
          "NOT_APPLICABLE"
        ],
        "title": "AnswerValue",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InitiativeCreate": {
        "properties": {
          "contact_email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "contact_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "organization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization"
          },
          "sector": {
            "title": "Sector",
            "type": "string"
          },
          "sector_other": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector Other"
          }
        },
        "required": [
          "name",
          "sector"
        ],
        "title": "InitiativeCreate",
        "type": "object"
      },
      "InitiativeRead": {
        "properties": {
          "contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "contact_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Name"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "organization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization"
          },
          "participant_type": {
            "title": "Participant Type",
            "type": "string"
          },
          "sector": {
            "title": "Sector",
            "type": "string"
          },
          "sector_other": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector Other"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "user_id",
          "name",
          "description",
          "sector",
          "sector_other",
          "contact_name",
          "contact_email",
          "organization",
          "participant_type",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "InitiativeRead",
        "type": "object"
      },
      "InitiativeStatus": {
        "enum": [
          "draft",
          "active",
          "submitted"
        ],
        "title": "InitiativeStatus",
        "type": "string"
      },
      "InitiativeUpdate": {
        "properties": {
          "contact_email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "contact_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "organization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "sector_other": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector Other"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InitiativeStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "InitiativeUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "contact": {
      "name": "CoE-DSC",
      "url": "https://coe-dsc.nl/"
    },
    "description": "API for the MAMI Framework DSI Assessment Tool \u2014 CoE-DSC / TNO",
    "title": "MAMI Checker API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/initiatives": {
      "get": {
        "description": "List all initiatives with owner email, status, and answer count.",
        "operationId": "list_initiatives_api_v1_admin_initiatives_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminInitiativeRow"
                  },
                  "title": "Response List Initiatives Api V1 Admin Initiatives Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Initiatives",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/initiatives/{initiative_id}": {
      "delete": {
        "description": "Delete an initiative and its child rows (answers, evidence, report), keeping the user.",
        "operationId": "delete_initiative_api_v1_admin_initiatives__initiative_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "initiative_id",
            "required": true,
            "schema": {
              "title": "Initiative Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Initiative",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/initiatives": {
      "post": {
        "operationId": "create_initiative_api_v1_initiatives_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiativeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiativeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Initiative",
        "tags": [
          "initiatives"
        ]
      }
    },
    "/api/v1/initiatives/me": {
      "get": {
        "operationId": "get_my_initiative_api_v1_initiatives_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiativeRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get My Initiative",
        "tags": [
          "initiatives"
        ]
      }
    },
    "/api/v1/initiatives/{initiative_id}": {
      "patch": {
        "operationId": "update_initiative_api_v1_initiatives__initiative_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "initiative_id",
            "required": true,
            "schema": {
              "title": "Initiative Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiativeUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiativeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Initiative",
        "tags": [
          "initiatives"
        ]
      }
    },
...
    "/api/v1/initiatives/{initiative_id}/submit": {
      "post": {
        "description": "Mark initiative as submitted. Idempotent \u2014 re-submitting an already-submitted one is OK.",
        "operationId": "submit_initiative_api_v1_initiatives__initiative_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "initiative_id",
            "required": true,
            "schema": {
              "title": "Initiative Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Submit Initiative",
        "tags": [
          "initiatives"
        ]
      }
    },
    "/api/v1/questionnaire/config": {
      "get": {
        "description": "Return the v2 questionnaire config for the current user's participant type (DSI or SP).",
        "operationId": "get_questionnaire_config_endpoint_api_v1_questionnaire_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Questionnaire Config Endpoint",
        "tags": [
          "questionnaire"
        ]
      }
    },
    "/api/v1/questionnaire/initiatives/{initiative_id}/answers": {
      "get": {
        "description": "Return all saved answers for an initiative (for save/resume \u2014 QUES-02).",
        "operationId": "get_answers_api_v1_questionnaire_initiatives__initiative_id__answers_get",
        "parameters": [
          {
            "in": "path",
            "name": "initiative_id",
            "required": true,
            "schema": {
              "title": "Initiative Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AnswerRead"
                  },
                  "title": "Response Get Answers Api V1 Questionnaire Initiatives  Initiative Id  Answers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Answers",
        "tags": [
          "questionnaire"
        ]
      }
    },
    "/api/v1/questionnaire/initiatives/{initiative_id}/answers/{question_id}": {
      "put": {
        "description": "Upsert one answer for a question. Creates on first save, updates on subsequent saves.\nEnforces ownership: current user must own the initiative.",
        "operationId": "upsert_answer_api_v1_questionnaire_initiatives__initiative_id__answers__question_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "initiative_id",
            "required": true,
            "schema": {
              "title": "Initiative Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Upsert Answer",
        "tags": [
          "questionnaire"
        ]
      }
    },
...
  }
}